+++ src/settings/suggesters/FileSuggester.ts
@@ -1,9 +1,15 @@
 // Credits go to Liam's Periodic Notes Plugin: https://github.com/liamcain/obsidian-periodic-notes
 
-import { TAbstractFile, TFile } from 'obsidian';
-import { TextInputSuggest } from './suggest';
+import { AbstractInputSuggest, App, TAbstractFile, TFile } from 'obsidian';
 
-export class FileSuggest extends TextInputSuggest<TFile> {
+export class FileSuggest extends AbstractInputSuggest<TFile> {
+  private inputEl: HTMLInputElement;
+
+  constructor(app: App, inputEl: HTMLInputElement) {
+    super(app, inputEl);
+    this.inputEl = inputEl;
+  }
+
   getSuggestions(inputStr: string): TFile[] {
     const abstractFiles = this.app.vault.getAllLoadedFiles();
     const files: TFile[] = [];
@@ -22,8 +28,8 @@ export class FileSuggest extends TextInputSuggest<TFile> {
     el.setText(file.path);
   }
 
-  selectSuggestion(file: TFile): void {
-    this.inputEl.value = file.path;
+  selectSuggestion(file: TFile, _evt: MouseEvent | KeyboardEvent): void {
+    this.setValue(file.path);
     this.inputEl.trigger('input');
     this.close();
   }