+++ src/settings/suggesters/FolderSuggester.ts
@@ -1,9 +1,15 @@
 // Credits go to Liam's Periodic Notes Plugin: https://github.com/liamcain/obsidian-periodic-notes
 
-import { TAbstractFile, TFolder } from 'obsidian';
-import { TextInputSuggest } from './suggest';
+import { AbstractInputSuggest, App, TAbstractFile, TFolder } from 'obsidian';
 
-export class FolderSuggest extends TextInputSuggest<TFolder> {
+export class FolderSuggest extends AbstractInputSuggest<TFolder> {
+  private inputEl: HTMLInputElement;
+
+  constructor(app: App, inputEl: HTMLInputElement) {
+    super(app, inputEl);
+    this.inputEl = inputEl;
+  }
+
   getSuggestions(inputStr: string): TFolder[] {
     const abstractFiles = this.app.vault.getAllLoadedFiles();
     const folders: TFolder[] = [];
@@ -22,8 +28,8 @@ export class FolderSuggest extends TextInputSuggest<TFolder> {
     el.setText(file.path);
   }
 
-  selectSuggestion(file: TFolder): void {
-    this.inputEl.value = file.path;
+  selectSuggestion(file: TFolder, _evt: MouseEvent | KeyboardEvent): void {
+    this.setValue(file.path);
     this.inputEl.trigger('input');
     this.close();
   }