+++ test/jest-setup.js
@@ -1,0 +1,3 @@
+// Set up Obsidian globals for the jest/jsdom test environment
+global.activeWindow = global.window || global;
+global.activeDocument = global.document;